Added Middeware to filter Business by location and category
