+++ server/routes/index.js
@@ -2,6 +2,7 @@
 import businessController from '../controller/business';
 import userController from '../controller/user';
 import reviewController from '../controller/review';
+import filterBy from '../middlewares/filter';
 
 module.exports = (app) => {
   app.get('/api', (req, res) => res.status(200).send({
@@ -11,7 +12,7 @@ module.exports = (app) => {
   app.post('/auth/signup', userController.signup);
   app.post('/auth/login', userController.singin);
 
-  app.get('/businesses', businessController.getBusinesses);
+  app.get('/businesses', filterBy.filter, businessController.getBusinesses);
   app.get('/businesses/:businessId', businessController.getBusiness);
   app.post('/businesses/', businessController.create);
   app.put('/businesses/:businessId', businessController.update);
